Display real slot display name in activity slot
