+++ src/pages/journal/components/activity-slot/activity-slot.html
@@ -10,7 +10,7 @@
       </div>
     </ion-col>
     <ion-col class="travel-title-column">
-      <h5 class="travel-title">Travel</h5>
+      <h5 class="travel-title">{{getTitle()}}</h5>
     </ion-col>
     <ion-col class="activity-code-column">
       <h5 class="activity-code">{{formatActivityCode(slot.activityCode)}}</h5>
