+++ src/providers/slot-selector/slot-selector.ts
@@ -3,7 +3,6 @@ import { SlotItem } from './slot-item';
 import { TestSlotComponent } from '../../pages/journal/components/test-slot/test-slot';
 import { TestSlot } from '../../pages/journal/journal.model';
 import { ActivitySlotComponent } from '../../pages/journal/components/activity-slot/activity-slot';
-import { Activity, activities } from './activity.constants';
 
 @Injectable()
 export class SlotSelectorProvider {
@@ -32,14 +31,6 @@ export class SlotSelectorProvider {
       default:
         return TestSlotComponent
     }
-  }
-
-  public activityCodeDisplayName(activityCode: string): string {
-    const matchingActivity: Activity = activities.find(a => a.activityCode === activityCode);
-    if (matchingActivity) {
-      return matchingActivity.displayName || matchingActivity.description;
-    }
-    return 'Unknown';
   }
 
   private isNonTestActivityCode(activityCode: string): boolean {
